Commiting fixed is_matched column to main after deleting my branch (easier than merging)
added extra code cells where I redo the merging process used before to get the is_matched column. Previously this was done on the answers_matched dataframe, but it only matched answers and emails on professionals_id which led to matching a lot of answers-questions-emails that were wrong since the answer was not by the person who received that email.

The fix was to add another column to match on which is the question_id. The correct dataframe that has the correct is_matched column is answers_matched